refactor(sip): Extract Class für 1000-Zeilen-Regel (SipCoreCallChannel, SipCallSignalingService)
Beide Dateien überschritten nach dem F011-Early-Media-Umbau + CF-066a-Merge die
Architektur-Grenze von 1000 Zeilen. Behoben per echter Extract Class (keine partial,
kein Baseline), verhaltensbewahrend:

- SipCoreCallChannel (1034→854): Medienparameter-Publikation in neue
  SipCallChannelMediaPublisher-Kollaboratorklasse ausgelagert (parse→enrich→policy→fire,
  once-per-session-Guard, Rekey-Signatur, Live-SRTP/DTLS-Keying-State). Channel reicht
  den Live-Kontext lazy über MediaPublishContext + Socket-Release/Fire/Terminate-Delegates
  durch — Laziness der Route-Probe exakt erhalten. Folgt dem bestehenden
  SipCallChannel*-Kollaborator-Muster.
- SipCallSignalingService (1015→963): Ingress-Response-Header-Builder in statische
  SipIngressResponseHeaders-Klasse (symmetrisch zu SipIngressRequestPolicy).

Verifikation: Core net9+net10 Release 0 Warnungen; Architektur 12/12; Core-Integration
1472/1472, 0 Skip.
